refactor: CCP-4535 rename group container
